Pressing g over the broom hands gets it.
p for pickup does the same.

Added the class Item. The player has a vector of items (an Inventory)
and i put a broom handle into main to test picking an item up.

From here, i could implement the ability to wield it and develop
fighting or write items on the map to be stored in a map-wide inventory.
A simple function, transfer, could be written to take an item from one
inventory to another.