Rewrote mapgen with Tiles, Range, Area, dig.
Tiles abstracts a grid as a simple char* with width and height. It simply
allows a clean way to access a one-dimensional array as a 2D one.

Range abstracts a range with a minimum and maximum. Area has a range for both
horizontal and verticle.

dig and dig_room take advantage of Range and Area to produce cleaner code than
add_room with random_rect (and Rect).